+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Copyright (c) 2021-2024 Claudio Satriano <satriano@ipgp.fr>
 
+## unreleased
+
+- New config options: `station_metadata_path` and `waveform_data_path` to
+  read station metadata and waveform data from files. Supports any metadata
+  format supported by ObsPy and SDS (SeisComp Data Structure) waveform
+  archives.
+
 ## v0.4.1 - 2024-03-11
 
 - Bugfix: `requake` executable was not installed